requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
sqlalchemy==2.0.15
alembic==1.11.1
pydantic==1.10.7
python-dotenv==0.19.0
celery==5.2.7
pytesseract==0.3.10
pillow==9.5.0
psycopg2-binary==2.9.5

